@@ -12,49 +12,48 @@
   },
   "private": true,
   "dependencies": {
-    "@angular/animations": "^6.0.6",
-    "@angular/common": "^6.0.6",
-    "@angular/compiler": "^6.0.6",
-    "@angular/core": "^6.0.6",
-    "@angular/forms": "^6.0.6",
-    "@angular/http": "^6.0.6",
-    "@angular/language-service": "^6.0.6",
-    "@angular/platform-browser": "^6.0.6",
-    "@angular/platform-browser-dynamic": "^6.0.6",
-    "@angular/router": "^6.0.6",
-    "angularfire2": "^5.0.0-rc.11",
-    "bootstrap": "^4.0.0",
-    "core-js": "^2.5.3",
-    "firebase": "^5.1.0",
-    "modernizr": "^3.6.0",
+    "@angular/animations": "~8.2.11",
+    "@angular/common": "~8.2.11",
+    "@angular/compiler": "~8.2.11",
+    "@angular/core": "~8.2.11",
+    "@angular/forms": "~8.2.11",
+    "@angular/platform-browser": "~8.2.11",
+    "@angular/platform-browser-dynamic": "~8.2.11",
+    "@angular/router": "~8.2.11",
+    "angularfire2": "^5.2.1",
+    "bootstrap": "^4.3.1",
+    "core-js": "^3.3.2",
+    "firebase": "^5.5.7",
+    "modernizr": "^3.7.1",
     "ngx-siema": "^2.0.1",
-    "ngx-toastr": "^8.8.0",
-    "rxjs": "^6.2.1",
-    "rxjs-compat": "^6.2.1",
-    "ts-node": "~5.0.1",
-    "typescript": "2.7.2",
-    "zone.js": "^0.8.26"
+    "ngx-toastr": "^11.1.2",
+    "rxjs": "^6.5.3",
+    "rxjs-compat": "^6.5.3",
+    "tslib": "^1.10.0",
+    "zone.js": "~0.9.1"
   },
   "devDependencies": {
-    "@angular/cli": "^6.0.8",
-    "@angular/compiler-cli": "^6.0.6",
-    "@types/dotenv": "^4.0.2",
-    "@types/jasmine": "~2.8.6",
-    "@types/jasminewd2": "~2.0.2",
-    "@types/node": "^9.4.5",
-    "@types/uuid": "^3.4.3",
-    "codelyzer": "~4.2.1",
-    "jasmine-core": "~3.1.0",
+    "@angular-devkit/build-angular": "~0.803.10",
+    "@angular/cli": "~8.3.10",
+    "@angular/compiler-cli": "~8.2.11",
+    "@angular/language-service": "~8.2.11",
+    "@types/dotenv": "^6.1.1",
+    "@types/jasmine": "^3.4.4",
+    "@types/jasminewd2": "^2.0.8",
+    "@types/node": "^12.11.1",
+    "@types/uuid": "^3.4.5",
+    "codelyzer": "^5.0.0",
+    "jasmine-core": "~3.4.0",
     "jasmine-spec-reporter": "~4.2.1",
-    "karma": "~2.0.0",
+    "karma": "~4.1.0",
     "karma-chrome-launcher": "~2.2.0",
-    "karma-cli": "~1.0.1",
-    "karma-coverage-istanbul-reporter": "^1.4.1",
-    "karma-jasmine": "^1.1.1",
-    "karma-jasmine-html-reporter": "^1.0.0",
-    "nodemon": "^1.14.12",
-    "protractor": "~5.3.0",
-    "tslint": "~5.9.1",
-    "@angular-devkit/build-angular": "~0.6.8"
+    "karma-coverage-istanbul-reporter": "~2.0.1",
+    "karma-jasmine": "~2.0.1",
+    "karma-jasmine-html-reporter": "^1.4.0",
+    "nodemon": "^1.19.4",
+    "protractor": "~5.4.0",
+    "ts-node": "~7.0.0",
+    "tslint": "~5.15.0",
+    "typescript": "~3.5.3"
   }
 }
