@@ -52,8 +52,8 @@ export class ProductsListComponent implements OnInit, OnDestroy {
 
   getProducts() {
     this.productsLoading = true;
-    this.productsCacheService
-      .get('products', this.productService.getProducts())
+    this.productService
+      .getProducts()
       .pipe(takeUntil(this.unsubscribe$))
       .subscribe((products) => {
         this.products = <Product[]>products;
